requirements.txt
psycopg2-binary
selenium
datetime